apps/py/requirements.txt
blinker==1.6.2
click==8.1.3
Flask==2.3.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
Werkzeug==2.3.2
requests==2.30.0
opentelemetry-exporter-jaeger-thrift==1.17.0
opentelemetry-sdk==1.17.0
opentelemetry-api==1.17.0
opentelemetry-instrumentation-flask==0.38b0
